Removed unit tests and also reorganized the files
