+++ Cargo.toml
@@ -16,8 +16,8 @@ license = "MIT"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+chrono = "0.4.28"
 clap = { version = "4.3.23", features = ["derive"] }
-tempfile = "3.8.0"
 term_size = "0.3.2"
 
 [profile.release]
