+++ src/args.rs
@@ -16,49 +16,3 @@ pub struct Cli {
     #[arg(short, long)]
     pub list: bool,
 }
-
-#[cfg(test)]
-mod tests {
-    use std::path::PathBuf;
-
-    use clap::Parser;
-
-    use crate::Cli;
-
-    #[test]
-    fn test_default_values() {
-        let cli = Cli::parse_from([""]);
-
-        assert_eq!(cli.path, None);
-        assert!(!cli.all);
-        assert!(!cli.list);
-    }
-
-    #[test]
-    fn test_parse_args() {
-        let args = ["", "--all"];
-        let cli: Cli = Cli::parse_from(args);
-
-        assert!(cli.all);
-        assert!(!cli.list);
-    }
-
-    #[test]
-    fn test_parse_path() {
-        let args = ["", "path/to/some/file"];
-        let cli: Cli = Cli::parse_from(args);
-
-        assert_eq!(cli.path, Some(PathBuf::from("path/to/some/file")));
-        assert!(!cli.all);
-        assert!(!cli.list);
-    }
-
-    #[test]
-    fn test_parse_list() {
-        let args = ["", "--list"];
-        let cli: Cli = Cli::parse_from(args);
-
-        assert!(cli.list);
-        assert!(!cli.all);
-    }
-}
